Reorganised file directories and made the app accessible via https
