+++ create.py
@@ -1,0 +1,4 @@
+from application import db
+
+db.drop_all()
+db.create_all()
